requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
pydantic==2.7.1
python-dotenv==1.0.0
requests==2.31.0
openai>=1.0.0
pytest==8.2.0
numpy<2
openenv-core>=0.2.0
 